Move calculateProcessDate to Salesforce object
